+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## ✨ Features
 
 * Upgrade cozy-bar@7.20.1 to be able to call onSelect function
+* Improve speed of search suggestion by preventing fetch notes url until click
 * Update cozy-stack-client and cozy-pouch-link to sync with cozy-client version
 * Update cozy-ui
   - Modify Viewers to handle [68.0.0 BC](https://github.com/cozy/cozy-ui/releases/tag/v68.0.0)